sdk/python/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "peerclaw-sdk"
version = "0.1.0"
description = "HTTP client for PeerClaw crews, flows, and tasks"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = ["httpx>=0.27", "pyyaml>=6"]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6", "mypy>=1.11"]

[tool.hatch.build.targets.wheel]
packages = ["src/peerclaw"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
